BR-NeXT/requirements.txt
# BR-NeXT core dependencies
numpy>=1.24
scipy>=1.10
jax>=0.4.20
jaxlib>=0.4.20
flax>=0.7.5
optax>=0.1.7
orbax-checkpoint>=0.4.0
onnx>=1.15
onnxruntime>=1.17
pyyaml>=6.0

# Optional: for direct ONNX export from JAX
# jax2onnx>=0.1.0

# Optional: CUDA backend (install matching your CUDA version)
# jax[cuda12_pip]>=0.4.20
